Setup/requirements.txt
stable-baselines3[extra]>=2.3.0
gymnasium>=0.29.1
torch>=2.2.0
mujoco>=3.2.0
wandb>=0.17.0
tensorboard>=2.15.0
imageio>=2.34.0
imageio-ffmpeg>=0.5.0
numpy>=1.26.0,<2.0.0
pyyaml>=6.0
tqdm>=4.66.0